Debugging but commiting now for safety

+++ src/test/java/io/github/greymagic27/jna_clone/WinDef/LongTest.java
@@ -9,14 +9,14 @@ class LongTest {
 
     @Test
     void testConstructorStoresValue() {
-        assertEquals(42, new LONG(42).longValue());
-        assertEquals(0, new LONG(0).longValue());
-        assertEquals(-123, new LONG(-123).longValue());
+        assertEquals(42, new LONG(42).intValue());
+        assertEquals(0, new LONG(0).intValue());
+        assertEquals(-123, new LONG(-123).intValue());
     }
 
     @Test
     void testLongValue() {
-        assertEquals(7L, new LONG(7).longValue());
+        assertEquals(7, new LONG(7).intValue());
     }
 
     @Test
